backend/requirements.txt
python-dotenv>=0.19.0
streamlit>=1.24.0
google-generativeai>=0.3.0
langchain>=0.1.0
langchain-google-genai>=0.0.5
python-docx>=0.8.11
chromadb>=0.4.14
protobuf>=3.20.0
langchain_community
langchain_openai
Flask
Flask-Cors
oauth2client
google-api-python-client
certifi
gunicorn==21.2.0  # or your preferred version
docx2pdf
python-docx
google-cloud-storage
Flask-SQLAlchemy
Flask-Migrate
psycopg2-binary
Flask-Bcrypt
PyJWT
blinker
Flask-JWT-Extended
twilio